Split out `getMimePrefix`
